Replace DiscoveryServerInfo (manual) with ServerDiscoveryInfo (generated) and move ServerVersion out of discovery package
