+++ jellyfin-core/api/jellyfin-core.api
@@ -8,8 +8,8 @@ public final class org/jellyfin/sdk/Jellyfin {
 }
 
 public final class org/jellyfin/sdk/Jellyfin$Companion {
-	public final fun getApiVersion ()Lorg/jellyfin/sdk/model/discovery/ServerVersion;
-	public final fun getRecommendedVersion ()Lorg/jellyfin/sdk/model/discovery/ServerVersion;
+	public final fun getApiVersion ()Lorg/jellyfin/sdk/model/ServerVersion;
+	public final fun getRecommendedVersion ()Lorg/jellyfin/sdk/model/ServerVersion;
 }
 
 public final class org/jellyfin/sdk/JellyfinOptions {
